requirements.txt
python-dotenv==1.1.0
google-generativeai==0.8.4
pandas==2.2.3
ortools==9.12.4544
matplotlib==3.10.1
openpyxl==3.1.5
pdm==2.23.0
uvicorn==0.32.0
fastapi==0.115.4
markdown==3.8